// src/app/models/cp-search-result.model.ts
export class CP_SearchResult {
  partnerId: string;
  firstName: string;
  middleName: string;
  lastName: string;
  fathersName: string;
  spousesName: string;
  occupation: string;
  chkFlag: string;
  updateNameFlag: string;
  updateFatherNameFlag: string;
  updateHusbandNameFlag: string;
  updateOccupationNameFlag: string;
  policyRef: string;
  contractStatus: string;
  roleType: string;
  sumAssured: number;
  startDate: Date;
  riskDate: Date;
  productId: string;
  ratedUp: string;
  reason: string;
  value: string;
  dateOfBirth: Date;
  gender: string;
  placeOfBirth: string;
  permanentAddress: string;
  telephone: string;
  mobile: string;
  confidenceParameter: number;
  numberOfPolicies: number;
  aadhaarNumber: string;
  panNumber: string;
  images: string[];

  constructor() {
    this.partnerId = '';
    this.firstName = '';
    this.middleName = '';
    this.lastName = '';
    this.fathersName = '';
    this.spousesName = '';
    this.occupation = '';
    this.chkFlag = 'N';
    this.updateNameFlag = '0';
    this.updateFatherNameFlag = '0';
    this.updateHusbandNameFlag = '0';
    this.updateOccupationNameFlag = '0';
    this.policyRef = '';
    this.contractStatus = '';
    this.roleType = '';
    this.sumAssured = 0;
    this.startDate = new Date();
    this.riskDate = new Date();
    this.productId = '';
    this.ratedUp = 'NO';
    this.reason = '';
    this.value = '';
    this.dateOfBirth = new Date();
    this.gender = '';
    this.placeOfBirth = '';
    this.permanentAddress = '';
    this.telephone = '';
    this.mobile = '';
    this.confidenceParameter = 0;
    this.numberOfPolicies = 0;
    this.aadhaarNumber = '';
    this.panNumber = '';
    this.images = [];
  }
}